Cleanup headers
