+++ templates/containers.h
@@ -25,12 +25,12 @@
 
 #include <vector>
 #include <unordered_map>
+#include "uhdm_types.h"
 
 #ifndef UHDM_CONTAINERS_H
 #define UHDM_CONTAINERS_H
 
 namespace UHDM {
-  typedef void any;
   typedef std::vector<any*> VectorOfany;
   <CONTAINERS>
 };
