+++ templates/group_header.cpp
@@ -32,7 +32,7 @@ namespace UHDM {
 
   bool <GROUPNAME>GroupCompliant(any* item) {
     BaseClass* the_item = (BaseClass*) item;
-    unsigned int uhdmtype = the_item->UhdmType();
+    UHDM_OBJECT_TYPE uhdmtype = the_item->UhdmType();
     if (<CHECKTYPE>) {
       std::cout << "Internal Error: adding wrong object type (" << UhdmName(uhdmtype) << ") in a <GROUPNAME> group!\n";   
       return false;
